Reorganize build, remove random module
