+++ .github/workflows/ci.yml
@@ -90,11 +90,11 @@ jobs:
 
       - name: Make target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: mkdir -p kernel-testkit/target random/target stats/target target .js/target kernel/.jvm/target kernel/.js/target laws/.js/target math/.js/target tests/js/target monte-carlo/target .jvm/target .native/target laws/.jvm/target tests/jvm/target core/target math/.jvm/target testkit/target project/target
+        run: mkdir -p kernel-testkit/target stats/target target .js/target kernel/.jvm/target kernel/.js/target laws/.js/target math/.js/target tests/js/target monte-carlo/target .jvm/target .native/target laws/.jvm/target tests/jvm/target core/target math/.jvm/target testkit/target project/target
 
       - name: Compress target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
-        run: tar cf targets.tar kernel-testkit/target random/target stats/target target .js/target kernel/.jvm/target kernel/.js/target laws/.js/target math/.js/target tests/js/target monte-carlo/target .jvm/target .native/target laws/.jvm/target tests/jvm/target core/target math/.jvm/target testkit/target project/target
+        run: tar cf targets.tar kernel-testkit/target stats/target target .js/target kernel/.jvm/target kernel/.js/target laws/.js/target math/.js/target tests/js/target monte-carlo/target .jvm/target .native/target laws/.jvm/target tests/jvm/target core/target math/.jvm/target testkit/target project/target
 
       - name: Upload target directories
         if: github.event_name != 'pull_request' && (startsWith(github.ref, 'refs/tags/v') || github.ref == 'refs/heads/main')
