+++ tests/jvm/src/test/scala/schrodinger/kernel/CategoricalSpec.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package schrodinger.random
+package schrodinger.kernel
 
 import cats.syntax.all.*
 import cats.instances.vector.*
